requirements/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
argon2-cffi==21.2.0
python-dotenv==0.21.0
redis==4.4.0
PyJWT==2.6.0
kavenegar==1.1.2
Pillow==9.3.0
unicode-slugify==0.1.5
django-filter
django-extensions==3.2.1
gunicorn==20.1.0
django-cors-headers==3.13.0
uvloop==0.17.0
httptools==0.5.0
psycopg2-binary==2.9.5
coverage==6.4.4
requests==2.28.1
freezegun==1.2.2